Run level mode changes
svn path=/branches/hk2-2-integration/; revision=54380


Former-commit-id: 2138d0daf49bdd2d452d96d1d93a12c8d9d19c78

+++ main/nucleus/core/kernel/src/test/java/com/sun/enterprise/v3/server/AppServerStartupTest.java
@@ -185,7 +185,7 @@ public class AppServerStartupTest {
         if (rla != null) {
             descriptorBuilder.to(RunLevel.class).
                     has(RunLevel.RUNLEVEL_VAL_META_TAG, Collections.singletonList(((Integer) rla.value()).toString())).
-                    has(RunLevel.RUNLEVEL_MODE_META_TAG, Collections.singletonList(rla.mode().toString()));
+                    has(RunLevel.RUNLEVEL_MODE_META_TAG, Collections.singletonList(((Integer) rla.mode()).toString()));
 
             descriptorBuilder.in(RunLevel.class);
         }
